requirements.txt
Faker==4.1.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.4
python-dateutil==2.8.1
scipy==1.4.1